Add login and signup logics

+++ Xplore/Models/Price.swift
@@ -1,5 +1,5 @@
 struct Price{
-    var pricePerDay : Int
+    var pricePerDay : Double
     var taxPercentage : Double
     var currencyCode : String
 }
